+++ src/rules/whitespaceRule.ts
@@ -132,6 +132,8 @@ class WhitespaceWalker extends Lint.SkippableTokenAwareRuleWalker {
                         prevTokenShouldBeFollowedByWhitespace = true;
                     }
                     break;
+                default:
+                    break;
             }
         });
     }