+++ src/tslint.ts
@@ -114,5 +114,7 @@ class Linter {
     }
 }
 
+// tslint:disable-next-line:no-namespace
 namespace Linter {}
+
 export = Linter;